% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{example_acyclic}
\alias{example_acyclic}
\title{Acyclic example}
\format{
Directed \link[igraph:aaa-igraph-package]{igraph::igraph} object with link dependencies encoded in
the \code{"synergy"} link attribute.
}
\source{
Wang, R.-S., Sun, Z. & Albert, R. (2013) Minimal functional routes in
directed graphs with dependent edges. \emph{Intl. Trans. in Op. Res.}, 20,
391--409.
}
\description{
This toy example, a directed acyclic graph, was used by Wang et al (2013) to
illustrate their constructions and algorithms.
}
